+++ CHANGELOG.md
@@ -1,14 +1,23 @@
 # Changelog
 
-## v0.0.21
+## v0.0.22
 
 ### Added minor updates
-  - `~/TextDoc.tsx`
+  - `~/deps.ts`
+  - `~/routes/_app.ts`
+
+### Updated the text component page with the new `compact` property
+  - `~/docs/atoms/text/compact.md`
+  - `~/docs/testTexts/aliceFirst.md`
+  - `~/routes/atoms/text.tsx`
+
+### Updated the TextDoc component look
+  - `~/components/TextDoc.tsx`
 
 ### Comming soon (v0.1.0)
 
 - Features
-  - [ ] Content driven documentation using Sanity.io or raw md files.
+  - [x] Content driven documentation using Sanity.io or raw md files.
     - [x] Add gfm to import richer .md content.
   - [x] Migrate testing pages from the ana-components repo